+++ src/js/components/counters.js
@@ -1,10 +1,9 @@
 import numberRollup from "@handmadeweb/number-rollup";
 
-let numberRolleups = document.querySelectorAll('.bmcb-counter');
+let numberRolleups = document.querySelectorAll('.bmcb-counter .counter-element');
 
 export default function () {
   if (numberRolleups.length) {
-    console.log('yes')
     numberRollup();
   }
 }